Refactor FileWatchService into separate file, making it less daunting
